Commit:4539b411cac560c33134390f8c21832d1176a4d3
Sample configuration file.

The boto.conf file contains the AWS authentication information
that boto requires to make a connection. While it is possible
to create boto.conf manually, it is unnecessary because you
will be prompted for you credentials on the first invocation of
quest.py and the boto.conf file will be created for you.